feat(ui): design at three widths — desktop, tablet, mobile; tighten density
The app had two ad-hoc breakpoints (1080, 820) plus four one-off ones
scattered through the stylesheet, and "tablet" just meant a narrow desktop.
There are now three deliberate tiers, documented in one block:

  Mobile   <=639px    off-canvas sidebar; tables restack as cards; modals
                      become bottom sheets; toolbars stack, with the
                      occasional contact actions folded into an overflow menu
  Tablet   640-1023   the sidebar collapses to a 68px ICON RAIL that stays on
                      screen — no hamburger, no overlay; two-up grids
  Desktop  >=1024     the base styles

The sidebar is now one NAV array rather than eight hand-written buttons, so
all three tiers render from the same definition and the labels can drop out
for the rail. Tables carry data-label on every cell, which is what lets them
restack; verified no horizontal overflow on any screen at any tier.

Density: the page frame, cards, tables, empty states and overlays were all
sized for a hero layout — 28px page padding around a 22px card inset, and a
78px medallion inside a 44px-padded empty state. Trimmed throughout, the
max content width raised from 1200 to 1560, and the logs card no longer
forces 500px of height when it holds one row.

Also flattened the remaining pill chrome (status labels, the header number,
the scan status) to match the badges and WhatsApp Web; round is kept only
where the shape is the meaning — avatars, dots, unread counts, progress.

Backend, found while screenshotting the dashboard:

- pg returns int8 as a string, so every COUNT(*) came back as text where
  SQLite gave a number. "Today's replies" rendered `0 + 0 + 1` as "001" and
  computed a response rate from it. int8 is now parsed as a Number in pgdb,
  which is safe at these magnitudes.
- SUM() returns numeric, which pg also stringifies and the parser above does
  not cover — the four response sums and the three scheduler summary sums are
  cast to int in SQL.
- 'sent' and 'cancelled' had no badge variant, so they rendered as bare text
  beside real badges.
- The activity log opened filtered to "Delivered", hiding the failures it is
  usually opened to find. It now opens on All.

+++ backend/scheduler.js
@@ -322,9 +322,9 @@ cron.schedule('* * * * *', () => {
             
             db.get(`
                 SELECT 
-                    SUM(CASE WHEN status IN ('delivered', 'read', 'sent') THEN 1 ELSE 0 END) as sentCount,
-                    SUM(CASE WHEN status = 'failed' THEN 1 ELSE 0 END) as failedCount,
-                    SUM(CASE WHEN status = 'pending' THEN 1 ELSE 0 END) as pendingCount,
+                    SUM(CASE WHEN status IN ('delivered', 'read', 'sent') THEN 1 ELSE 0 END)::int as sentCount,
+                    SUM(CASE WHEN status = 'failed' THEN 1 ELSE 0 END)::int as failedCount,
+                    SUM(CASE WHEN status = 'pending' THEN 1 ELSE 0 END)::int as pendingCount,
                     MIN(sent_time) as firstSentTime
                 FROM automation_logs 
                 WHERE automation_id = ? AND sent_time >= ? AND sent_time <= ?
